requirements.txt
praw==5.0.1
python-telegram-bot==6.1.0
logging==0.4.9.6